rename file to all lowercase
